Adding clean data sql sheet
In this step, the focus is on cleaning the data to ensure it is in the correct format and free from inconsistencies. Since this dataset is a Google public dataset, it is relatively clean compared to real-world client data. However, some cleaning was still necessary, particularly with timestamp data formats.